% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaUI.R
\name{arimaUI}
\alias{arimaUI}
\title{ARIMA parameter picker}
\usage{
arimaUI(ts.ar)
}
\arguments{
\item{ts.ar}{A time series}
}
\value{
The function open a viewer panel.
It allows to choose the arima parameters.
The value returned is a list with the arima parameter and the coefficients of a fit of the time serie.
}
\description{
Gadget allowing to choose easily the parameter of an ARIMA model.
Some help to the decision making process is displayed as wikipedia link
}
\note{
The Ljung-Box tab panel come from an article of http://alstatr.blogspot.co.uk/.
It had been kindly made available by @alstated.
}
\examples{
sim.ts <- arima.sim(
n = 500,
model = list(
  order = c(1, 0, 2),
  ar = 0.8
, ma = c(0.2, 0.5)
))

list.res <- arimaUI(sim.ts)

# comparison of manual selection with automatic one
library(forecast)
auto.arima(sim.ts)

}

